form validation outsourced as hook
